+++ tests/unit/navigator/test_navigation_scenario.py
@@ -15,7 +15,7 @@ class TestNavigationScenario(TestCase):
         self.firmware = Firmware.NANOS
         self.callbacks = dict()
         self.navigator = MagicMock()
-        self.navigate_with_scenario = NavigateWithScenario(self.navigator, self.firmware.device,
+        self.navigate_with_scenario = NavigateWithScenario(self.navigator, self.firmware,
                                                            "test_name", self.directory)
 
     def tearDown(self):